+++ scout/server/blueprints/variants/controllers.py
@@ -1,7 +1,7 @@
 import decimal
 import logging
 import re
-from typing import Any, Dict
+from typing import Any, Dict, Optional
 
 import bson
 from flask import Response, flash, session, url_for
@@ -16,20 +16,22 @@ from scout.adapter import MongoAdapter
 from scout.constants import (
     ACMG_COMPLETE_MAP,
     ACMG_MAP,
+    CANCER_EXPORT_HEADER,
     CANCER_SPECIFIC_VARIANT_DISMISS_OPTIONS,
     CANCER_TIER_OPTIONS,
     CHROMOSOMES,
     CHROMOSOMES_38,
     CLINSIG_MAP,
     DISMISS_VARIANT_OPTIONS,
+    EXPORT_HEADER,
     EXPORTED_VARIANTS_LIMIT,
+    FUSION_EXPORT_HEADER,
     MANUAL_RANK_OPTIONS,
     MOSAICISM_OPTIONS,
     SPIDEX_HUMAN,
     VARIANT_FILTERS,
     VARIANTS_TARGET_FROM_CATEGORY,
 )
-from scout.constants.variants_export import CANCER_EXPORT_HEADER, EXPORT_HEADER
 from scout.server.blueprints.variant.utils import (
     callers,
     clinsig_human,
@@ -1024,21 +1026,17 @@ def download_str_variants(case_obj, variant_objs):
     )
 
 
-def download_variants(store, case_obj, variant_objs):
+def download_variants(
+    store: MongoAdapter, case_obj: dict, variant_objs: CursorType, category: Optional[str] = None
+) -> Response:
     """Download filtered variants for a case to an excel file
-
-    Args:
-        store(adapter.MongoAdapter)
-        case_obj(dict)
-        variant_objs(PyMongo cursor)
 
     Returns:
         an HTTP response containing a csv file
     """
-    document_header = variants_export_header(case_obj)
-    export_lines = []
+    document_header = variants_export_header(case_obj, category)
     export_lines = variant_export_lines(
-        store, case_obj, variant_objs.limit(EXPORTED_VARIANTS_LIMIT)
+        store, case_obj, variant_objs.limit(EXPORTED_VARIANTS_LIMIT), category
     )
 
     def generate(header, lines):
@@ -1060,7 +1058,9 @@ def download_variants(store, case_obj, variant_objs):
     )
 
 
-def variant_export_lines(store, case_obj, variants_query):
+def variant_export_lines(
+    store: MongoAdapter, case_obj: dict, variants_query: CursorType, category: Optional[str] = None
+):
     """Get variants info to be exported to file, one list (line) per variant.
     Args:
         store(scout.adapter.MongoAdapter)
@@ -1084,7 +1084,7 @@ def variant_export_lines(store, case_obj, variants_query):
         variant_line.append("_".join([str(position), change]))
 
         # gather gene info:
-        gene_list = variant.get("genes")  # this is a list of gene objects
+        gene_list = variant.get("genes", [])  # this is a list of gene objects
 
         # if variant is in genes
         if gene_list:
@@ -1092,10 +1092,10 @@ def variant_export_lines(store, case_obj, variants_query):
             variant_line += gene_info
         else:
             empty_col = 0
-            while empty_col < 4:
+            while empty_col < 5:
                 variant_line.append(
                     "-"
-                )  # empty HGNC id, empty gene name and empty transcripts columns
+                )  # empty HGNC id, empty gene name, two empty transcripts columns, and consequence
                 empty_col += 1
 
         if variant.get("cadd_score"):
@@ -1108,7 +1108,35 @@ def variant_export_lines(store, case_obj, variants_query):
         else:
             variant_line.append("N/A")
 
-        if case_obj.get("track") == "cancer":
+        if category == "fusion":
+            for field in ["fusion_genes", "orientation", "frame_status", "found_db"]:
+                value = variant.get(field, "N/A")
+                if "," in value:
+                    value = value.split(",")
+                if isinstance(value, list):
+                    value = " | ".join(value)
+                if value is None or value == []:
+                    value = "N/A"
+                variant_line.append(value)
+
+            exon = ""
+            for gene in gene_list:
+                for transcript in gene.get("transcripts", []):
+                    if "exon" in transcript:
+                        if exon:
+                            exon = exon + " | "
+                        exon = exon + transcript["exon"]
+            variant_line.append(exon)
+
+            variant_gts = variant["samples"]  # list of coverage and gt calls for case samples
+            for individual in case_obj["individuals"]:
+                for variant_gt in variant_gts:
+                    if individual["individual_id"] != variant_gt["sample_id"]:
+                        continue
+                    variant_line.append(variant_gt.get("read_depth", "N/A"))
+                    variant_line.append(variant_gt.get("split_read", "N/A"))
+                    variant_line.append(variant_gt.get("ffpm", "N/A"))
+        elif case_obj.get("track") == "cancer":
             # Add cancer and normal VAFs
             for sample in ["tumor", "normal"]:
                 allele = variant.get(sample)
@@ -1224,7 +1252,7 @@ def variant_export_genes_info(store, gene_list, genome_build="37"):
     return gene_info
 
 
-def variants_export_header(case_obj):
+def variants_export_header(case_obj: dict, category: str = "snv") -> list:
     """Returns a header for the CSV file with the filtered variants to be exported.
     Args:
         case_obj(scout.models.Case)
@@ -1233,7 +1261,9 @@ def variants_export_header(case_obj):
                 + AD_reference, AD_alternate, GT_quality for each sample analysed for a case
     """
     header = []
-    if case_obj.get("track") == "cancer":
+    if category == "fusion":
+        header = header + FUSION_EXPORT_HEADER
+    elif case_obj.get("track") == "cancer":
         header = header + CANCER_EXPORT_HEADER
     else:
         header = header + EXPORT_HEADER